Remove neg from proposal.
These were included to make proposal a superset of label. But they don't make sense in proposals. A label is issued as the result of voting on a proposal, and a negation would be the result of a changed vote or new vote on the same proposal. But the proposal itself can't be negated.

Further, I see no reason why a proposal should expire. Proposals should be a permanent record -- even if there are time limits involved in the process of voting on proposals.